+++ libraries/linear/tcc/BiMatrix.tcc
@@ -4,7 +4,7 @@ namespace linear
 {
 
     template<typename RowMatrixType>
-    const typename BiMatrix<RowMatrixType>::RowType& BiMatrix<RowMatrixType>::GetColumn(uint index) const
+    const typename BiMatrix<RowMatrixType>::RowType& BiMatrix<RowMatrixType>::GetColumn(uint64 index) const
     {
         return _transp.GetRow(index);
     }
